Changed date field parsing from text based solution template
The date field is now read from the text based solution templates
and used as such.

This tries to mimic reading the json or yaml template data even
though the text based template intention is a bit different. The
text templates are intended to be used to update and process
the solutions and snippets. The yaml and json are meant for data
migration.

It could be considered that in case of text template, the import
operation could be treated as create or update in which case the
content date stamp should be updated to current time regardless
of what user set.

Also the behaviour of importing text template and processing
the date field differs depending on if the data was imported
with or without message digest. In case of import without
specifying the message digest, the date is read from the
template. But in case the data is updated with message
digest, the date gets updated based on current time stamp.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>