Fixed default solutions to match latest templates
The previously saved default solutions did not have the
single trailing newline. The logic that make sure that
there is only one new line at the end of solution was only
recently implemented.

One of the solutions also missed the DATE field in the
header.

The solution content data and the text header DATE field
is not updated when the solution is updated. Only the
content metadata gets updated automatically. The thought
is that user may want to control the visible date. The
tool tracks the latest change and updates the content
date metadata automatically.

Signe-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>