Changed instantiating of Config()
The Config() is not created anymore in Snippy __init__. This
tries to create functionality for testing purposes where the
sys.argv is not used when the Snippy() is instantiated. This
allows doing the steps like below which is considered more
logical for test case reading.

    $ snippy = Snippy()
    $ sys.argv = ['snippy', 'import', '--solution', '--defaults']
    $ cause = snippy.run_cli()

Also in the future there might be something else than the CLI
interface to run. For example REST might be implemented and in
that case the sys.argv cannot is not used.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>