Fixed workflow tests with Python2.7
New refactoring of workflow tests to create isolated tests
where for examples the Snippy() is created for each test
created a situation in Pyhon2.7 that the sqlite3 database
was still shared.

The problem was that the Sqlite in Python2.7 does not
support in memory database. This is done so that there
is a file for the database in Python2. This file and the
database was still shared between the new tests which made
them fail.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>