Changed solution delete operation internals
Same change was made for solution than for snippet in previous
commit.

The implementations for solution and snippet are quite close
each others but it is currently though that separating the
implementation into two different modules helps with
readability and maintainability. If both content categories
would be combined, there would be likely small differences
which would gradually degrade the code simplicity.

Added more tests for snippet and solution delete and export
operations.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>