Added idea to refactor migrate operations
The migrate export of content could be made in similar
manner than delete and update. In this case there could
be n-amount of content to be dumped.

Add Config.is_search_context() and run the search in
similar manner than with update and delete. The first
option is always the template and last option migrate
everything.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>