Fixed snippy.reset() not to create new Config object
The Snippy reset() method was creating new Config() object
when Config().reset() was called and when the Config().reset
returned. This was previously added because sequential tests
failed if one test failed. The reason was that previously
the same Snippy() object was used to run multiple tests and
console commands.

Currently all the workflow tests create new Snippy() object
and this is not currently problem.

It is not known at this point of time if the reset() really
does what it is planned because it is not really used. Also
in some cases when one test fails, it may fail few (but not
all) following tests. This is likely to be due to other
issues.

The removal of the new Config() from two places dropped the
case run time from ~2.5 seconds to ~2.1 seconds in a fast
machine so the impact was very large.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>