Fixed running individual tests
The previous commit that changed the Snippy.reset() was
faulty. It created a situation (and revealed fault) where
the reset() is called before run_cli(). This created a
situation where the Config.args was not assigned with
Arguments() which lead to exception.

There is still the thinking error or resetting Arguments()
which is left unsolved in this commit. That is, it does
not make sense to reset Arguments. Just the configuration
needs to be recreated or reset to allow multiple operations
with the same object.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>