Changed the way how Config() is initialized
The intention is to be able to use Snippy also with
REST API. The intention is that the configuration
can be initialized from multiple sources. Also the
intention is that the configuration remains for the
whole active session of Snippy.

The session concept is to help testing and to keep
the in-memory database active when multiple CLI
commands must be run to create different scenarios
from storage point of view. From normal usage point
of view, single run should be immediately released.

The Config() is now initialized with source like
Arguments(). In the future, the REST API needs to
provide another class to implement same API as the
Arguments() object and pass it to Config().

Removed the Snippy.reset() because it was not a good
idea. The test cases require that the cause is reset
because multiple commands are run with the same
session. Because of this, the cause is reset at the
end of run_cli().

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>