Fixed exit cause printing
Previous commit for 'Changed the way how Config() is
initialized' broke the exit cause printing. The cause
was always printed as 'OK' also in case of failure.

Test cases did not detect this because the check was
made from the snippy.run_cli() return value that was
correct but not from stdout which was not correct.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>