Fixed cause for updating or deleting with search
If search options like --sall or --sgrp were used with delete
or update, the tool exit with OK cause even when the content
operation was not executed. The fault was in case where the
search option matched to more than one content. In this case
the operation is always prevented since the search option for
update and delete must be always exactly one content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>