Fixed user errors for create content without changes
Following fixes apply on to content create operation.

If the solution template was not modified at all and
it was still saved from editor, the solution content
was incorrectly stored. Now this generates proper error
cause.

Added proper error cause if the solution template was
not used at all and the solution input was just an
empty string.

In case of snippet content, the data is always parsed
so that the content data is empty if it was missing.
Because of this, there is no reason to compare against
empty template usage for snippet content.

Added test cases to verify the behavior.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>