Fixed content category identification
The content category must now be explicitly identified when
importing and creating content with editor. Previously the
solution category was selected even if the content was not
identified.

The thought is to be more strict about the content being
read in order to avoid storing data in database that is
not what is wanted and what cannot be properly parsed.

Fixed also the text template content reading to read
with read() instead of readlines(). The text template was
read with read() that resulted list that was passed to
Editor() as edited parameter. The editor always needs
the file in string. Now the vi editor and text file based
input methods pass the file as a string with newlines.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>