Fixed failing workflow test for help commands
The command line parser uses argparse that exits after
options like --help. This causes SystemExit but this cannot
mocked because if mocked (in straightforward manner), the
exit is not happening and it seems that the argparse is
(for some reason) run multiple times with empty namespace.
In Python2.7 this is three times and with Python3 this is
two times.

If the exit is tested 'with self.assertRaises(SystemExit)',
it blocks the assert tests to validate the given output.
Most likely the SystemExit is bypassed and test case
failures are not visible at all. The SystemExit allows
checking the exit cause but nothing else.

The current way is to catch the SystemExit exception
and run the asserts then. This is currently the only
known way that works in Python2, Python3 and allows
checking stdout and run assets.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>