Added test to print test reference
The '--help tests' allows printing test cases to find out
what is tested and what is not that is now tested.

The reference documentation file reading was changed. The
reason was that the file mock does not seem to support
iterators so doing next from the infile seems not to be
possible. Because of this, the implementation was changed
to read the file into list and process the test case files
line by line from a list.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>