Added new test (skipped) to verify console logging
There is now a test that tries to verify that the console
logs are printed with '-vv' option.

This test is skipped at this point of time.

For some reason running this test with all the test slows
down all the other tests. There was similar problem few
commits ago. At that point of time it was related to order
where the sys.argv is defined for the test. The problems
starts if the sys.argv is before the creation of Snippy()
object. In this case this would be a must since the logging
object is created at the initialization phase.

More investigation needed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>