Added Logger() reset to set default logging level
The previously described CPU peaks and major slowness while
running all tests if the sys.argv is defined before the
Snippy() object is created seems to be related to usage
of debug logs. The problem was at least visible when -vv
or --debug options were used. It seemed that the usage of
these options left the log level even when the Snippy was
terminated with None.

It seems that there were too many logs and the pytest tries
to sort them based on timestamps which seemed to increase
the CPU and test case execution time.

The pytest has -vv option as well and the Logger and
Profiler classes are reading debug options directly from
sys.argv. Not sure if this is related or not.

But I cannot understand why the log level was left on
debug level and cases started to fail if the sys.argv
was specified before the Snippy() was created? The log
level should reset when the object is cleaned? But does
this mean that something is left hanging?

In order to clear the situation and run tests for debug
options, the Logger now has reset() which is called in
Snippy.release().

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>