Changed make test to not to use -vv option
There is a fault in snippy tool because the pytest -vv
option affects to snippy. There are few places where
the option is read directly from sys.argv. These are
the --debug, -vv and --profile to get the values before
the Arguments() class has been initialized.

For now the tests are run without the -vv option. If
this option is enabled, the performance test and the
'no stderr' check fails.

TODO:

  1. Is it possible to make the Logger so that it
     would somehow see this and prevent? How?

  2. The another question is that why some logs are
     sent to stderr sometimes?

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>