Fixed error cause for missing search criteria
If user did not add any search criteria like, keyword, digest
or data, the error cause indicated internal error that was
not true. For example command 'snippy search' caused the
situation.

Now the error text assumes that there is no internal error
and user did not set any of the search criteria.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>