Changed sqlite3 cursor usage
The sqlite3 cursor is now automatically closed after each
operation with context manager.

There is a bit conflicting or hard to understand information
whether the cursor really needs to be closed or not. For
example /1/ says to close it and /2/ that it is not necessary.

But as of now, it is considered with context manager and
immediate closing of the cursor is better approach like
explained in /1/.

/1/ https://www.youtube.com/watch?v=D7wSMnapDp4
/2/ https://stackoverflow.com/a/22618781

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>