Added metadata to exported YAML and JSON files
The exported YAML and JSON data now contains metadata that
contains for example tool version that was used to create
the file. The intention is to be able identify easily the
format of the data and if there are database schema changes
between different versions.

The information is not yet used for any purpose by the
tool.

The metadata is not created for text file exports. Text
files were considered error prone to parse with additional
data that is not actual content data. At least as of now.
The YAML and JSON are intended for content migration where
the data is moved between users and tools.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>