Changed error log to error cause
The error log was replaced with error cause which is displayed
to the user when tool exists. The idea is to reduce glutter
and not to print any logs on the screen when the tool is run
There is only OK or NOK with a cause at the end. If more
information is needed, it is always possible to enable more
logging with -vv or --debug options.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>