Removed unused constant
The content columns do not contain 'testing' column anymore.
This was previously used in the first version of the testing
framework but it was removed.

Fixed 'magic number' in test framework that was presenting
the total number of columns in the content. Added new constant
that better describes the value instead of 'magical number'.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>