Removed the Storage.debug() method
The database debug method was not used and it was removed.
A better way to troubleshoot storage problems is the database
helper that can be used with a test case to print database
content.

Also it was not a good idea to try to dump whole database
at the end. This did not even work work because the Sqlite
connection was released before the database was tried to
be dumped.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>