Changed all logs to be printed to stdout
All logs with one exception are now always printed to stdout.
The exception is the version number that is printed with the
help of argparse. The problem is that Python changed the way
how the print command sends the logs to stdout or stderr.
Previously the default for Python and argparse was stderr
before Python 3.4. From Python 3.4 and onwards the default
is stdout.

The argparse own implementation to print the version should
be changed to own version printing method to print the version
always to stdout. This is not yet done.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>