Fixed Sqlite3 database initialization
In case the Sqlite3 connection already exists, the database
is not initialized again. The original approach for the tool
was to have a single run execution from command line command.
Now quite many tests the same Snippy() object is used to
run multiple operations.

For the possible REST server mode, this is also better
approach if the server is constantly running.

There was small performance gain on this change because
this prevented the reinitialization of the database.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>