Fixed import failures when running tests
Importing the Gunicorn and Falcon was always made when
tests were executed. The intention is not yet to force
installing these two dependencies since the REST API
is in early development phase.

The import errors for these two dependencies are ignored
as of now since the --server option is not meant to be
used.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>