Fixed REST responses JSON format
The JSON standard mandates usage of double quotes in JSON
strings. The /hello API was using single quoted JSON which
was incorrect. The reason Falcon has the response.media
variable that is translated to JSON. This is new in Falcon
1.3.

Based on /1/ "Falcon's Internet media type defaults to
application/json but you can set it to whatever you like."
This seems to support yaml and MessagePack.

/1/ https://falcon.readthedocs.io/en/stable/user/tutorial.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>