Changed Config() method calls to use class name
The Travis CI is failing with core dump in Python 3.6. The
root cause is not known. There seems to be a second exception
when the console tests are made and the SystemExit exception
is handled. This second exception seems to cause some cases
to fail in Python 3.6 with Travis CI. This is not happening
on personal PC.

It is not sure if this fixes the Travis CI core.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>