Fixed unhandled exception in command line argument parsing
The Cli() and the argparse exit with SystemExit in case
support option like --help is used or the parsing fails.
This caused the Config() not to read the source config
properly. This caused second exception because the
internal data structures for the Config() were not set.

The design is now that the argparse can exit in case
of failures or with special options. This does not
execute the Cli processing logic currently to the end
of method. This is considered bit "iffy" but it is
easier and more straightforward as of now. If argparse
would not exit or it would be captured, the error
handling would have to be improved in run_cli() method
and it would be more complicated. This is considered
more straightforward way to go.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>