Removed execution of one test case to debug core
The Travis CI is causing cores with Python 3.6. One of the
snippet delete workflow tests fails. One of them was removed
it is not sure if this was the failing one because the test
case names are not printed since the -vv option in pylint
caused the logs to be printed in Snippy. This option is
read by Snippy directly from the command line.

Added also the /hello API test case back to see how it
goes in Travis CI.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>