Fixed test failures with Pytest 3.3
There is a new feature in Pytest 3.3 that has something
to do with logging. This seems to capture logs from the
Python logger itself.

In this case there is likely an incorrect usage of root
logger in snippy/logger/logger.py when the logs are
disbled.

This is not fully understood at this point of time but
the workaround in /1/ helps.

The problem appears only with Pytest 3.3.

/1/
https://github.com/pytest-dev/pytest/issues/2982

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>