Fixed sqlite database content leak between tests
The new REST API cases used to create the Snippy() instance
but they did not release it. This seemed to leave sometimes
(randomly?) database content hanging from previous test.

The storage is passed into REST server and stored there.
The variable is not released between tests if it is not
explicitly released.

The exact problem and why the problem behaves like this
is not fully understood. The best understanding is that
since the variable holding the database reference is not
released, it can persist some content at some point of
time over tests.

It cannot be ruled out that there would not be other
problems related to this as well.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>