Fixed 'snippy --help tests' command
The command was failing when executed once but this failure
was not visible when all tests were executed.

The reason is the realization that static class variables
are shared between every instance of Snippy() in the same
process. This causes situation where the Config() was
initialized in test test_wf_console_help when the subtest
test_console_help_tests was executed. When the case is
executed individually, the Config() is not initialized and
it behaves as expected. But when tests are run, there is
always multiple instances from Snippy() running from the
same test case file. This caused the unexpected leak of
configuration between tests.

This is now fixed in snippy.release() that resets the
configuration. This is also documented as (accidental)
design principle.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>