Fixed Snippy logger not to use root logger
The Logger() is now using the 'snippy' logger like:

  logging.getLogger('snippy').disabled = True

This also fixes the failure with the latest Pytest 3.3
that was circumvented with parameter '-p no:logging' for
Pytest. More information in commit c827b32b4be2.

The problem was likely because the Snippy and Pylint both
used the root logger and there were some consequences that
are not fully understood as of now. But now the no:logging
parameter is not needed anymore for the Pytest.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>