Fixed leaking logger handlers
Previously the Logger() class generated always new handler
even when the previous handler for the same module already
existed. This cumulated multiple logs from the same log.
This was especially a problem in tests that run several
instances from Snippy() in the same module. The problem
originated for example from static class variable usage
for logger. For example Config() used to behave like
this.

With the latest changes the usage of -vv option for pytest
is not a problem anymore. Also the -p no:logging is not
needed. The reason was the previous commit that changed
the Logger() to use 'snippy' logger and not the root
logger.

How ever, this change caused some side effects to console
tests that capture the stdout. For some reason the test
'test_console_very_verbose_option' is failing when run
with all other tests. When the test is run individually
it works. This indicates probably similar problems that
with the Config() where the same instance is shared with
all Snippy() objects in the tests. Also the failing test
seem to print the data to pytest stdout but for some reason
the sys.stdout is not capturing this. (TODO)

The performance gain was so good that this change is taken
into use even though there may be some smaller bugs still
to solve.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>