Fixed exit cause when run without server modules
There is now proper failure message if Snippy was
installed without server modules Gunicorn and Falcon
when the tool was tried to be run in server mode.

The module checks from server specific modules were
removed. Now the check is only in one place where the
server is started. Also the exit cause in case of
missing modules is now properly set.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>