Fixed logging hierarchy
Removed unnecessary 'snippy' hierarchy level because the
snippy was twice in the logger hierarchy names.

The 'CRITICAL' log level was set explicitly for the tool
highest log hierarchy 'snippy'. Before this, the level was
'NOTSET' which caused the log level to be inherited from
root level. If Snippy() would be used with other modules,
the root level change would then affect to Snippy() as
well. Now this is prevented and changing the root level
log level does not have affect to Snippy().

The logging still relies that only the highest log
hierarchy sets the level and all the levels below it don't.
change the 'NOTSET' level. This allows easy change of log
levels that affects to all levels by just changing the
highest hierarchy level log level.

Added also debug method to print the logging hierarchy
with the new Python module logging_tree. This is installed
only in development mode and is not intended to be used
with official installation.

There is still the failure with console workflow tests
and the 'test_console_very_verbose_option'

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>