Fixed capturing logger output
The console help workflow test now relies on Pytest caplog
method. This requires that parameter '-p no:logging' is
_not_ used. This parameter disables the log capturing of
logging output from Pytest and it fails the test that
relies on caplog.

The problem is that it seems that Pytest does not allow
capturing the logs from stdout even though the Snippy.
logger sets the stream handler to stdout. For reason that
is not understood at the moment the trick to mock sys.
stdout with StringIO does not work. The logs are visible
only from Pytest.

The test that uses the logcap does not use inherited
class 'unittest.TestCase'. This is because the use of
logcap requires Pytest and the test class cannot
inherit from unittest in this case.

This was tested with Pytest 3.3.1 which is not required
explicitly by the setup.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>