Fixed --sgrp search option in case 'match all'
The search option --sgrp was not able to print all the
content if the 'match all' was used. In case the search
keyword is dot, empty or not used at all, the result
must be 'match all'. This same behaviour is applied also
to --sall and --stag. The 'match all' prints all the
content in this case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>