Fixed pytest version in Docker test suites
There is the one console help test that relies on pytest
caplog fixture which requires recent Pytest. The Pytest
version is required to be latest one available as of now.

Removed also explicit downgrades for py and pluggy Python
modules. These were matched against the ones that Travis
CI uses with Python 3.4. But latest fixes are working
and these are running nicely in Travis and it was
considered that using latest or default is better.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>