Fixed multilevel sorting in GET /snippets&sort
The multilevel sorting based on multiple columns is now
relying on logic where the order of user provided columns
is first reverted. Then the sort is run for each column
one by one. This is based on example and comment in /1/.

This seems to work at least with two levels in one test.
But whether this is 100% failure proof, cannot be stated
at this point of time.

More testing is needed.

/1/ https://stackoverflow.com/a/4233482 (@tomcounsell)

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>