Added cause setting for REST API
There is now initial implementation for cause code setting
that operates with HTTP status codes. The intention is that
this will replace the previous text based cause setting.

The new HTTP based cause code operates on JSON structures
and is able to push multiple failures in the cause list.
This provides more information about the actual failure.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>