Added POST /api/snippets for REST API
It is now possible to create new snippets with the JSON
REST API.

It was noticed that the reference performance test was
degraded quite much. The reference test of 55 iteration
loop is down from 0.78 to 1.15. It was quickly tested
that this is not related to searching and returning
the snippet content for REST API in search() and
create(). # TODO

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>