Changed reading caller module and code line
The call the inspect.stack() was optimized. This call caused
file access for each frame that was expensive. The stack(0)
had pretty good performance bump but the sys._getframe was
slightly better. See /1/ and /2/ for more information.

The problem with the current optimization is the note that
the _getframe 'It is not guaranteed to exist in all
implementations of Python.' This is referring to CPython
which means that this may not work there. The performance
gain compared to stack(0) might not eventually justify
this change after all. Currently this change is added with
'let's see what happens' mentality.

/1/ https://stackoverflow.com/a/17407257
/2/ https://stackoverflow.com/a/45196608

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>