Fixed tests after adding versioning into REST API
The version number was added to the REST API. The base URL
is now /api/v1. It was stated that the nature of a REST
API is static and it should not change. But it is likely
that all the issues can be sorted with the first try and
it was considered better to add the versioning.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>