Changed --links option internals
The -l|--links is now internally behaving like a list. The
behaviour now looks a bit like the -t|--tags option. This was
preparing work for the REST API and link list processing.

This did not change the external behaviour. The links must be
still separated with space from the command line.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>