Fixed link (URL) separation characters
Previously comma was allowed to separate links. The comma
is a safe character in URL character set and thus it can be
used in link (URL) /1/.

Now links in a string context can be separated with space or
bar. This also means that the tool does no parse links with
space or bar characters correctly. If these two characters
are encoded as requested /1/, they are parsed correctly.

In JSON REST API, the links and tags can be given in list
context in a JSON data structure normally.

/1/ https://perishablepress.com/stop-using-unsafe-characters-in-urls/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>