Fixed OAS specifications to define character set
The OpenAPI specifications were updated to correctly define
the consumed and produced characters sets as UTF-8.

Author cannot claim to fully understand the character set
encodings and it is likely that other than UTF-8 formats
may be consumed. It is likely that used REST API framework
converts different sets to UTF-8. The produced character
encoding is UTF-8 because this is coming from Falcon FW.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>