Fixed OpenAPI specifications and content types
In case of DELETE and GET, the requests and/or responses should
(or MUST NOT) have body. Previously the content type was defined
for these operation as JSON. Now the content types are excluded
in case of DELETE and GET operations if the body is not present.

This also caused definition and correction of error responses.
Now the error responses and defined correctly. In case of some
responses, the body will be there. For example DELETE with 404
will provide body that contains the errors.

The Snippy server does not process a body in GET and DELETE
requests. The response of GET contains error or list of content.
The response for successful delete does not contain body. In
case of delete failure, the body contains list of errors.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>