Fixed GET and search keyword parameters
In case of search queries were made through REST API with
sall, stag or sgrp with one keyword, the parameter was not
parsed correctly. The problem was that the command line
always returned a list but the REST API with one parameter
resulted string.

Now all parameters that are needed in list context from
configuration source, are resulted as list.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>