Changed OpenAPI specifications for PUT and POST
The specifications now require using list context in case
new content is POSTed or updated with PUT. Also the responses
which result for example the created content result now always
list in the OpenAPI specifications.

It is not yet sure if this makes sense. The intention is to
harmonized the behaviour so that always a list would be
processed. This could allow updating or creating multiple
contents with the same request.

This is to be meditated more.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>