Fixed REST API content data in list context
It is now possible to give the content data from REST API
in list context so that each line is one element in a list.

The tool enforces only one newline at the end of each line.
That is, the lines may or may not have newline or have
multiple newlines when sent in the POST request. The content
data lines will be always stored with one newline.

The OAS specifications define the content data in list
context but it also works in one string where each line
contains newlines as pleased.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>