Added partial support to add multiple snippets
There is now partial support to add multiple snippets at once
with the POST /api/v1/snippets. The tests are made with one
item in the snippet list (in the request body) and there is
at least one know issues with setting the cause code. If there
are multiple snippets inserted at once, the OK cause is not
returned correctly because there will be multiple causes in
the list. Also there is a possibility to have different
failures and this may be error prone and hard to handle.

It was also realized that the list of snippets does not make
sense with the PUT when the snippet is identified with a digest
in path. In this case the OpenAPI specifications were changed
so that the snippet is not in a list in this case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>