Changed cause code setting in success cases
The cause code setting was optimized in success cases. In case
the cause is one of the OK HTTP causes, the called module and
line number from calling point are not set. The reason is that
setting the line number is expensive because it requires
reading the code file.

This gives performance boost mainly with single iteration
delay dropped from ~0.0365 to ~0.0237 which is significant.
The longer the run is (or repeating the same OK causes) the
less this matters. Most likely due to some kind of caching
of the files when constantly read.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>