Added API reference performance test
The test is slow and the reason is not known at the moment.
Slowness is linear in sense that the delay is not increased
if iterations are increased. This should indicate that there
are no leaks. The current guess is that this is due to Python
requests module that creates the connections always per each
requests.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>