Changed API reference performance test
Using http.client (or httplib in Python 2) will give about
20% performance boost. But the API code is still 4 times
slower than the CLI code. With Python request module the
API code was 5-6 times slower.

With http.client the performance results do not have so
much jitter.

The reason seems to be that the http.client is a lower
lever library than the Python requests. This explains
the performance boost in here.

Fixed also the Python 2 regression which was failing
because httplib had changed the name in Python3.

Next the Snippy code needs to be profiled.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>