Added programming language classifiers
The packaging configuration now has programming language
classifiers for Python 2 and Python 3. It was not really
sure if these are needed or how the behaviour changes
because the dot releases of support Python versions were
already listed. This syntax with the base Python 2 and 3
releases was used in the official Python packaging example
/1/ so it was copied here too.

/1/ https://packaging.python.org/tutorials/distributing-packages/

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>