Fixed configuration source list default value
The default value for list was defined as string when it
should have been empty list. This was causing list with
single element that was an empty string when the list
should have been empty.

Changed configuration base logging

Defined rules how to write logs and changed the config
source base class to use them.

Changed method names and code comments

Renamed two method names in config source base and in
config to better match what the methods were doing.

Fixed content dictionary generation

Fixed the content dictionary generation so that it can
handle removal of dictionary keys that are not in the
content dictionary. There was a special case with the
'key' that is the database primary key that is never
sent to end user. Previously this was handled in a bit
awkward way where the 'key' was always removed from the
removed field list. Now the error handling is improved
to allow removal of non-existent dictionary keys from
the content.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>