Fixed Python 2 unicode string conversion
Some of the parameters can be read in list of string
context. In Python 3 the string is always str that is
always a unicode string. But in Python 2, a string can
be str or unicode string. The conversion was missing
for methods taking care of the sorted and removed
fields.

Changed the Python 2 conversion code to own method.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>