Fixed empty list in REST API responses
If links or tags were not defined for the content, the
REST API JSON contained a list that had one element
which was an empty string. The list must be an empty
list in this case.

If the links or tags were not defined for the content,
the REST API JSON contained a list that had one element
which was an empty string. The list must be an empty
list in this case.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>