Fixed setup.cfg referring to non-existent file
The setup.cfg was referring to README.md file that does not
exist anymore. Also the file set the description-file that
was not used from the setup.py. The purpose of the setup.cfg
was clarified a bit and it makes sense to keep the file.

The reason to keep the file as of now is to define the
universal wheel to support Python 2 and Python 3. Also there
might be security risks with the setup.py and running code
from it /1/.

The plan is also to add the description file from README.rst
like in /2/ once the security considerations are checked.

/1/ https://stackoverflow.com/a/45108506
/2/ https://github.com/pypa/sampleproject/blob/master/setup.py

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>